Avoid showing notifications when an unsupported game is active
It seems to be common for users to run qbms operations and not wait
for them to complete. In the meantime they may change to a different
game mode, at which point they may receive qbms notifications and may
erroneously assume that it's the currently active game reporting these.

This extension REALLY needs re-writing to function as a module and not
an event based system.